+++ src/types/internal.ts
@@ -192,6 +192,7 @@ export interface ITestOptions {
     getExpectedStatus: (context: ITestSettingValueGetterContext) => Promise<number>;
     getHeaders: (context: ITestSettingValueGetterContext) => Promise<Record<string, any>>;
     getParameters: (context: ITestSettingValueGetterContext) => Promise<Record<string, string>>;
+    getQuery: (context: ITestSettingValueGetterContext) => Promise<Record<string, string>>;
     getTimeout: (context: ITestSettingValueGetterContext) => Promise<number>;
     index: number;
     method: Function;